requirements.txt
fastapi
uvicorn[standard]
pdf2image
Pillow
google-genai
python-multipart
Jinja2
markdown
PyMuPDF